requirements.txt
Flask==2.3.3
Pillow==10.0.0
PyJWT==2.8.0
natsort==8.4.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Werkzeug==2.3.7
uuid==1.30
python-dotenv==1.0.0
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
requests==2.31.0 